+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 * Rebound `projectile-compile-project` to <kbd>C-c p c</kbd>
 * Reworked `projectile-compile-project` and `projectile-test-project` to be smarter, more configurable and closer in behavior to the stock `compile` command
+* `projectile-switch-project` now runs `projectile-find-file` instead of `dired`
 
 ## 0.9.1 (04/26/2013)